% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomsplitviolin.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Split Violin Plot}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes} or \code{aes_}.}

\item{data}{The data to be displayed in this layer. It can be omitted, in which case
the default data defined in the ggplot object will be used.}

\item{...}{Additional arguments passed to \code{layer}. These are often aesthetics,
used to set an aesthetic to a fixed value, like \code{color = "red"} or \code{size = 3}.
They may also be parameters to the paired geom/stat.}

\item{draw_quantiles}{Numeric vector of quantiles to draw as horizontal lines
inside the violins.}

\item{trim}{If \code{FALSE}, doesn't trim the tails of the violins to the range of the data.}

\item{scale}{Scale parameter for the violin plot.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning.
If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them.}
}
\value{
A ggplot layer
}
\description{
Creates a split violin plot, allowing for the visualization of the distribution
of a variable for two groups side-by-side in the same violin plot.
}
\examples{
ggplot(data = YourDataFrame, aes(x = factorVariable, y = numericVariable)) +
  geom_split_violin()
}
